+++ endpoint/endpoint_test.go
@@ -1,4 +1,4 @@
-package actions
+package endpoint
 
 import (
 	"errors"
